+++ test/test.js
@@ -1,18 +1,5 @@
 'use strict';
 
-try {
-    console.log;
-}
-catch (error) {
-    window.console = {
-        log: function () {
-            var pre = document.createElement('pre');
-            pre.append(Array.join(arguments, ', '));
-            document.body.append(pre);
-        }
-    };
-}
-
 lib.dom.ready().then(function () {
 
     var windows = [],